// frontend/src/gen/akashic/v1/passage_pb.ts
export class Passage extends Message<Passage> {
  /**
   * @generated from field: int32 id = 1;
   */
  id = 0;

  /**
   * @generated from field: int32 bank_id = 2;
   */
  bankId = 0;

  /**
   * @generated from field: int32 category_id = 3;
   */
  categoryId = 0;

  /**
   * @generated from field: string title = 4;
   */
  title = "";

  /**
   * @generated from field: repeated akashic.v1.PassageParagraph paragraphs = 5;
   */
  paragraphs: PassageParagraph[] = [];

  /**
   * @generated from field: akashic.v1.Difficulty difficulty = 6;
   */
  difficulty = Difficulty.UNSPECIFIED;

  /**
   * @generated from field: google.protobuf.Timestamp created_at = 7;
   */
  createdAt?: Timestamp;

  /**
   * @generated from field: google.protobuf.Timestamp updated_at = 8;
   */
  updatedAt?: Timestamp;

  constructor(data?: PartialMessage<Passage>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "akashic.v1.Passage";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "id", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
    { no: 2, name: "bank_id", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
    { no: 3, name: "category_id", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
    { no: 4, name: "title", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 5, name: "paragraphs", kind: "message", T: PassageParagraph, repeated: true },
    { no: 6, name: "difficulty", kind: "enum", T: proto3.getEnumType(Difficulty) },
    { no: 7, name: "created_at", kind: "message", T: Timestamp },
    { no: 8, name: "updated_at", kind: "message", T: Timestamp },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Passage {
    return new Passage().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Passage {
    return new Passage().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Passage {
    return new Passage().fromJsonString(jsonString, options);
  }

  static equals(a: Passage | PlainMessage<Passage> | undefined, b: Passage | PlainMessage<Passage> | undefined): boolean {
    return proto3.util.equals(Passage, a, b);
  }
}